requirements.txt
agno>=1.1.8
openai>=1.12.0
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
rich>=13.7.0
streamlit>=1.24.0
plotly>=5.13.0
textblob>=0.17.1
requests>=2.31.0
python-dotenv>=1.0.0
